agent/pyproject.toml
[tool.poetry]
name = "jarvis-agent"
version = "0.1.0"
description = ""
authors = ["MuzykaAndrii <andrymyzik@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
livekit-agents = "^0.12.17"
livekit-plugins-google = "^0.11.1"
livekit-plugins-cartesia = "^0.4.11"
livekit-plugins-deepgram = "^0.7.1"
livekit-plugins-silero = "^0.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
